Move forecast chart utils to its own file
