+++ src/store/dashboard/awsOcpDashboard/awsOcpDashboardWidgets.ts
@@ -2,11 +2,11 @@ import { ForecastPathsType, ForecastType } from 'api/forecasts/forecast';
 import { ReportPathsType, ReportType } from 'api/reports/report';
 import messages from 'locales/messages';
 import {
-  ComputedForecastItemType,
   ComputedReportItemType,
   ComputedReportItemValueType,
   DatumType,
 } from 'routes/views/components/charts/common/chartDatum';
+import { ComputedForecastItemType } from 'routes/views/components/charts/common/chartDatumForecast';
 import { awsDashboardWidgets } from 'store/dashboard/awsDashboard';
 import { DashboardChartType } from 'store/dashboard/common/dashboardCommon';
 import { formatCurrency, formatUnits } from 'utils/format';
