+++ src/store/dashboard/ociDashboard/ociDashboardWidgets.ts
@@ -3,11 +3,11 @@ import { ReportPathsType, ReportType } from 'api/reports/report';
 import messages from 'locales/messages';
 import { routes } from 'routes';
 import {
-  ComputedForecastItemType,
   ComputedReportItemType,
   ComputedReportItemValueType,
   DatumType,
 } from 'routes/views/components/charts/common/chartDatum';
+import { ComputedForecastItemType } from 'routes/views/components/charts/common/chartDatumForecast';
 import { DashboardChartType } from 'store/dashboard/common/dashboardCommon';
 import { formatUnits } from 'utils/format';
 import { formatPath } from 'utils/paths';
