+++ src/store/dashboard/ocpDashboard/ocpDashboardWidgets.ts
@@ -4,11 +4,11 @@ import { RosPathsType, RosType } from 'api/ros/ros';
 import messages from 'locales/messages';
 import { routes } from 'routes';
 import {
-  ComputedForecastItemType,
   ComputedReportItemType,
   ComputedReportItemValueType,
   DatumType,
 } from 'routes/views/components/charts/common/chartDatum';
+import { ComputedForecastItemType } from 'routes/views/components/charts/common/chartDatumForecast';
 import { DashboardChartType } from 'store/dashboard/common/dashboardCommon';
 import { formatCurrency, formatUnits } from 'utils/format';
 import { formatPath } from 'utils/paths';
